# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: WhatsApp Desktop Backend API
  version: 1.0.0
  description: |
    Cross-platform desktop application for WhatsApp messaging and automation using official Business Cloud API.
    Phase 1: Authentication & Single Send
    Phase 2: Campaigns & Scheduling
    Phase 3: Inbox & Automation (Current)
    Phase 4: Reports & Packaging
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
paths:
...
  /contacts:
    get:
      summary: List contacts
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  total:
                    type: integer

    post:
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
  /contacts/import:
    post:
      summary: Import contacts from CSV/XLSX
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
...
  /business-hours:
    get:
      summary: List business hours configuration (Phase 3)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessHours'

    post:
      summary: Create or update business hours entry (Phase 3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessHoursCreate'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessHours'
  /business-hours/is-open:
    get:
      summary: Check if currently within business hours (Phase 3)
      parameters:
        - name: timezone
          in: query
          schema: { type: string, default: "UTC" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  isOpen:
                    type: boolean
                  currentTime:
                    type: string
                  timezone:
                    type: string
  /business-hours/{dayOfWeek}:
    put:
      summary: Update business hours for a specific day (Phase 3)
      parameters:
        - name: dayOfWeek
          in: path
          required: true
          description: Day of week (0=Sunday, 6=Saturday)
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessHoursUpdate'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessHours'

    delete:
      summary: Delete business hours for a specific day (Phase 3)
      parameters:
        - name: dayOfWeek
          in: path
          required: true
          description: Day of week (0=Sunday, 6=Saturday)
          schema: { type: integer }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contact:
      type: object
      properties:
        id: { type: string }
        phoneE164: { type: string }
        name: { type: string }
        email: { type: string }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
    ContactCreate:
      type: object
      properties:
        phoneE164: { type: string }
        name: { type: string }
        email: { type: string }
        tags:
          type: array
          items: { type: string }
      required: [phoneE164, name]
...
    BusinessHours:
      type: object
      properties:
        id: { type: string }
        timezone: { type: string }
        day_of_week: { type: integer, description: "0=Sunday to 6=Saturday" }
        start_time: { type: string, format: time, description: "HH:MM:SS" }
        end_time: { type: string, format: time, description: "HH:MM:SS" }
        created_at: { type: string, format: date-time }
    BusinessHoursCreate:
      type: object
      properties:
        timezone: { type: string }
        day_of_week: { type: integer }
        start_time: { type: string }
        end_time: { type: string }
      required: [timezone, day_of_week, start_time, end_time]
    BusinessHoursUpdate:
      type: object
      properties:
        timezone: { type: string }
        start_time: { type: string }
        end_time: { type: string }
...